requirements.txt
streamlit
langchain
openai
PyPDF2
python-dotenv
